+ Added edit forms for category controller.

+++ Northwind.Web/Middleware/PageNumberMiddleware.cs
@@ -10,7 +10,7 @@ namespace Northwind.Web.Middleware
 {
     public class PageNumberService
     {
-        public short pageSize;
+        public int pageSize;
         public PageNumberService(IConfiguration config)
         {
             pageSize = Int16.Parse(config.GetSection("section0:M").Value);
